feat(mobile): android threads inbox and host connection state
