+++ apps/mobile/src/components/host/host-connection-state.tsx
@@ -1,25 +1,12 @@
-import { Button, Host, Image, ProgressView, Text, VStack } from '@expo/ui/swift-ui';
-import {
-  buttonStyle,
-  font,
-  multilineTextAlignment,
-  textSelection,
-} from '@expo/ui/swift-ui/modifiers';
-import { FOOTNOTE, SECONDARY } from '@mobile/components/form/styles.ios';
+import type { HostConnectionStateProps } from '@mobile/components/host/host-connection-state.types';
+import { Button, useThemeColor } from 'heroui-native';
+import { WifiOffIcon } from 'lucide-react-native';
+import { ActivityIndicator, Text, View } from 'react-native';
 import { useTranslations } from 'use-intl';
 
-const CENTERED = multilineTextAlignment('center');
-const TITLE = font({ textStyle: 'title2', weight: 'semibold' });
-
-export interface HostConnectionStateProps {
-  status: 'connecting' | 'error';
-  url: string;
-  /** The underlying failure, when the controller reported one. */
-  failure?: string;
-  onRetry: () => void;
-}
-
-/** Full-screen fallback shown while a host connection is being established or has failed. */
+/** Android full-screen fallback while a host connection is being established or has failed.
+ * Deliberately plain RN: the failure text stays selectable (`selectable`), which Compose text
+ * cannot offer, and nothing here is list chrome that would want MD3 dress. */
 export function HostConnectionState({
   status,
   url,
@@ -27,38 +14,36 @@ export function HostConnectionState({
   onRetry,
 }: HostConnectionStateProps): React.ReactNode {
   const t = useTranslations('mobile.connection');
+  const muted = useThemeColor('muted');
 
   return (
-    <Host style={{ flex: 1 }} useViewportSizeMeasurement>
-      <VStack spacing={16}>
-        {status === 'connecting' ? (
-          <>
-            <ProgressView />
-            <Text modifiers={[SECONDARY]}>{t('connecting')}</Text>
-          </>
-        ) : (
-          <>
-            <Image systemName="wifi.exclamationmark" size={44} modifiers={[SECONDARY]} />
-            <VStack spacing={6}>
-              <Text modifiers={[TITLE, CENTERED]}>{t('unavailableTitle')}</Text>
-              <Text modifiers={[SECONDARY, CENTERED, textSelection(true)]}>
-                {t('error', { url })}
-              </Text>
-            </VStack>
-            <Button
-              label={t('retry')}
-              systemImage="arrow.clockwise"
-              modifiers={[buttonStyle('borderedProminent')]}
-              onPress={onRetry}
-            />
-            {failure ? (
-              <Text modifiers={[FOOTNOTE, SECONDARY, CENTERED, textSelection(true)]}>
-                {failure}
-              </Text>
-            ) : null}
-          </>
-        )}
-      </VStack>
-    </Host>
+    <View className="flex-1 items-center justify-center gap-4 px-6">
+      {status === 'connecting' ? (
+        <>
+          <ActivityIndicator />
+          <Text className="text-muted">{t('connecting')}</Text>
+        </>
+      ) : (
+        <>
+          <WifiOffIcon size={44} color={muted} strokeWidth={1.5} />
+          <View className="items-center gap-1.5">
+            <Text className="text-center font-semibold text-foreground text-title">
+              {t('unavailableTitle')}
+            </Text>
+            <Text selectable className="text-center text-muted">
+              {t('error', { url })}
+            </Text>
+          </View>
+          <Button onPress={onRetry}>
+            <Button.Label>{t('retry')}</Button.Label>
+          </Button>
+          {failure ? (
+            <Text selectable className="text-center text-footnote text-muted">
+              {failure}
+            </Text>
+          ) : null}
+        </>
+      )}
+    </View>
   );
 }
