+++ apps/mobile/src/components/host/host-connection-state.types.ts
@@ -1,0 +1,7 @@
+export interface HostConnectionStateProps {
+  status: 'connecting' | 'error';
+  url: string;
+  /** The underlying failure, when the controller reported one. */
+  failure?: string;
+  onRetry: () => void;
+}
