+++ apps/mobile/src/components/host/thread-list/thread-list.tsx
@@ -1,65 +1,75 @@
-import { List, Section } from '@expo/ui/swift-ui';
-import { listStyle, refreshable } from '@expo/ui/swift-ui/modifiers';
-import type { SessionInfo } from '@linkcode/schema';
-import type { ThreadGroup } from '@linkcode/ui/native';
-import { useEffect, useState } from 'react';
+import {
+  Host,
+  LazyColumn,
+  PullToRefreshBox,
+  Row,
+  Spacer,
+  Text,
+  useMaterialColors,
+} from '@expo/ui/jetpack-compose';
+import { clickable, fillMaxWidth, padding, weight } from '@expo/ui/jetpack-compose/modifiers';
+import { Fragment, useState } from 'react';
+import type { ThreadListProps } from './thread-list.types';
 import { ThreadRow } from './thread-row';
+import { useThreadListState } from './use-thread-list-state';
 
-const MINUTE = 60000;
-
-/** The thread inbox body: one collapsible section per group. Grouping is decided by the
- *  caller — this only renders it. Collapsed keys live here because the state is presentational:
- *  nothing outside the list cares which groups are open.
- *
- *  `sidebar` is not cosmetic — SwiftUI only honours a `Section`'s expanded state under that list
- *  style, so it is what makes the groups collapsible at all. */
+/** Android thread inbox body. Compose has no collapsible Section: each group renders a clickable
+ * MD3 subheader row and conditionally its rows. PullToRefreshBox takes a controlled flag instead
+ * of awaiting the promise, so the spinner state is adapted here. */
 export function ThreadList({
   groups,
   labelFor,
   onOpenThread,
   onRefresh,
-}: {
-  groups: ThreadGroup[];
-  labelFor: (group: ThreadGroup) => string;
-  onOpenThread: (sessionId: SessionInfo['sessionId']) => void;
-  /** Drives SwiftUI's own pull-to-refresh, which holds its spinner until this resolves. */
-  onRefresh: () => Promise<void>;
-}): React.ReactNode {
-  const [collapsed, setCollapsed] = useState<ReadonlySet<string>>(() => new Set());
-  const [now, setNow] = useState(Date.now);
+}: ThreadListProps): React.ReactNode {
+  const { collapsed, setGroupExpanded, now } = useThreadListState();
+  const colors = useMaterialColors();
+  const [refreshing, setRefreshing] = useState(false);
 
-  useEffect(() => {
-    const timer = setInterval(() => setNow(Date.now()), MINUTE);
-    return () => clearInterval(timer);
-  }, []);
-
-  const setGroupExpanded = (key: string, expanded: boolean) =>
-    setCollapsed((current) => {
-      const next = new Set(current);
-      if (expanded) next.delete(key);
-      else next.add(key);
-      return next;
-    });
+  const refresh = () => {
+    setRefreshing(true);
+    void onRefresh().finally(() => setRefreshing(false));
+  };
 
   return (
-    <List modifiers={[listStyle('sidebar'), refreshable(onRefresh)]}>
-      {groups.map((group) => (
-        <Section
-          key={group.key}
-          title={labelFor(group)}
-          isExpanded={!collapsed.has(group.key)}
-          onIsExpandedChange={(expanded) => setGroupExpanded(group.key, expanded)}
-        >
-          {group.sessions.map((session) => (
-            <ThreadRow
-              key={session.sessionId}
-              session={session}
-              now={now}
-              onPress={() => onOpenThread(session.sessionId)}
-            />
-          ))}
-        </Section>
-      ))}
-    </List>
+    <Host style={{ flex: 1 }} useViewportSizeMeasurement>
+      <PullToRefreshBox isRefreshing={refreshing} onRefresh={refresh}>
+        <LazyColumn contentPadding={{ top: 4, bottom: 24 }} modifiers={[fillMaxWidth()]}>
+          {groups.map((group) => {
+            const expanded = !collapsed.has(group.key);
+            return (
+              <Fragment key={group.key}>
+                <Row
+                  verticalAlignment="center"
+                  modifiers={[
+                    clickable(() => setGroupExpanded(group.key, !expanded)),
+                    fillMaxWidth(),
+                    padding(16, 14, 16, 4),
+                  ]}
+                >
+                  <Text style={{ typography: 'titleSmall' }} color={colors.primary}>
+                    {labelFor(group)}
+                  </Text>
+                  <Spacer modifiers={[weight(1)]} />
+                  <Text style={{ typography: 'labelMedium' }} color={colors.onSurfaceVariant}>
+                    {expanded ? '▾' : '▸'}
+                  </Text>
+                </Row>
+                {expanded
+                  ? group.sessions.map((session) => (
+                      <ThreadRow
+                        key={session.sessionId}
+                        session={session}
+                        now={now}
+                        onPress={() => onOpenThread(session.sessionId)}
+                      />
+                    ))
+                  : null}
+              </Fragment>
+            );
+          })}
+        </LazyColumn>
+      </PullToRefreshBox>
+    </Host>
   );
 }
